Replace usage of "evaluate" in parsing with regex-check
Corner case when calling e.g. "f(x)". Expression is successfully parsed
by the parser, but will fail when assigning vectors to "f" and "x".
Solution was to assign ones to varaibles and evaluate. This failed
in situations as x/(z-y) -> 1/0.
Logic is replaced by usage of regex string split and validating
functions/operators in front if "(".