Improvised UX for editedData prop. Now the layers will by default point to the selectedWells, selectedDrawing properties in editedData props. So now if user want to pass editedData prop it can be passed without any property editedData: {}
Also, no need to declare selectedWells: "@@#editedData.selectedWells" in the layers prop.
Same for selectedPie property.

For drawingLayer, data and selectedFeatureIndexes is passed by default, hence no need to specify it explicitly.
If user want to preload data in drawingLayer it can be passed using those data and selectedFeatureIndexes properties.