chore: group some dependencies for dependabot (#2061)
The default behavior of dependabot is to generate one PR for each
dependency that needs to be updated. This leads to multiple PRs for
related dependencies, while related dependencies should be handled
together. For example storybook and 7 @storybook/* packages have the
same version number, and should be handled as a whole.

This PRs groups some dependencies:
- react
- deckgl
- storybook