fix: need to reset twice to go back (#2151) (#2173)
When the reset button is pressed, the default zoom is set. This causes
the scrolling to change and triggers the onScroll event, which sets the
overall zoom. This problem was solved by setting the "isDefZoom" flag,
which is set to "true" when setting the default scale and does not allow
the onScroll event to change the zoom

Co-authored-by: DShamakin <denis.shamakin@emerson.com>
Co-authored-by: t0oF <93762994+w1nklr@users.noreply.github.com>