fix: "[NGRM]: zooming using mouse scroll wheel is limited #2240" (#2260)
Will fix limited zoom in 2D and also missing ticks on the ruler top left
corner.

For large UTM like numbers there is some precision issues in 2D . Hence
will limit zoom somewhat in this case. Is a bit hacky but seems to work
fine.

A better possible future solution would be to translate all data to
around origo. For readout the numbers would have to be translated back.
This is similar to how RMS does it.