fix: Improve sort by stratigraphy depth in utility-lib for well-completions-plot (#2321)
Improve sorting of wells when sorting by stratigraphy depth.
- Improve sorting to look at all zones with open completions when
sorting.
- Update unit tests

---
**Summary:**

The sorting now creates an array of zone indices with open completion.
This array is sorted with increasing indices. Comparison function then
looks on element by element and compares the zone index for each well.
The array with lowest zone index (i.e. shallowest zone) will be placed
first in ascending sort, or last for descending sort. Continues as long
the zone index values are equal. If common part is equal and one list is
longer than the other, it means it has more completions - thus placed
first for ascending order, or last for descending order. Otherwise
arrays are equal.