fix: Update title after setting tracks (#2384)
Sometimes title is missing in the WellLogView. This occurs because while
track data is still loading from server, the tracks have incomplete data
so an empty log is fed to the component which has no well name. finally
when data arrives from server, the component is updated and tracks are
fed to the component, however, the title is not updated.
To fix this, when setting tracks to controller, the title (if it is not
a specific string or div element) should be updated with the well name
from the updated tracks.

Co-authored-by: t0oF <93762994+w1nklr@users.noreply.github.com>