feat: Fixes: [NGRM]: Slow readout caused by high pick depth when displaying readout data. (#2407)
Consider the following webviz react property that controls the readout
popup:

    /**
     * Parameters for the InfoCard component
     */
    coords?: {
        visible?: boolean | null;
        multiPicking?: boolean | null; 
        pickDepth?: number | null;
    };
The pickDepth has a default value of 10. This slows down the readout
considerably for large sets of data like for instance a big map.

When hovering over object in the view, pickDepth larger than one will
cause the readout to also display values for objects hidden behind the
closest.

I'm am changing the default value to 2. This will remedy the slow
readout and will be enough to show readout for for instance a well and a
map behind it. Or 2 wells on top of each other. If one wants more
pickDepth one will explicitly have to set this property.