fix: "74736: PPA: well trajectory artifacts" (#2463)
There are some artifacts when drawing wells with many nodes/legs.

This is caused by deck.gl not being able to handle to many nodes/legs in
a well path. Also there is z-buffer fighting between two sub layers in
the well layer implementation.

The solution is to resample and reduce number of nodes in a well path in
such a manner that its shape is retained.
This is done automatically in the well layer except when the standard
"dataTransform" property is set in which case the user is responsible of
any transform done to the path via this property.

A storybook example will be provided with an example of "dataTransform"
property usage.

---------

Co-authored-by: Havard Bjerke <havard.bjerke@aspentech.com>