feat: Add tooltip optional prop to track template (#2488)
It is now possible to customize the tooltip that appear when mouse
hovers on top of the title of the track. Instead of showing the title of
the track in the tooltip it is now possible to supply a custom tooltip
in the track template, that replaces the default text.


![image](https://github.com/user-attachments/assets/30c0f05f-da0f-4d7a-8ed4-1b8919c4088c)