feat: enhanced well label styling (#2487)
Continuation of #2433

Fixes #767 and #947 

Summary of changes:
* @nilscb added functionality for positioning and backgrounds for well
labels
* Extracted a new deck.gl layer `WellLabelLayer` which is used to render
labels in the `WellsLayer`. Its props can be specified using the
`wellLabel` prop in `WellsLayer`
* In order to prevent z buffer fighting, z coordinates were modified by
@nilscb. I modified it slightly so that a single layer can display
consistently in both Orthographic and Orbit views. I also narrowed the
near and far plane scope for Orthographic views in order to further
reduce z buffer collision.
* Created some stories to demonstrate the styling and positioning
options for well labels
* Added procedurally generated synthetic wells in order to demonstrate
the above


![image](https://github.com/user-attachments/assets/00bef14d-c479-45ce-bd8e-36f39c1faf4c)

---------

Co-authored-by: Nils Christian Bonnevie <nils.christian.bonnevie@gmail.com>