feat: add seismic layer
The seismic layer does display the cage of the seismic data and seismic slices.
The API is "generic" as it allows to specify a set of triangles with their tex coordinates. The target is to display seismic slices
(axis aligned) or fences (arbitrary sections).
Seismic values are given as a 2D array of values. The values are converted to a 2D texture and shaders are used to apply a colormap.

The seismic layer extends a CompositeLayer, and uses a PrivatePolylinesLayer for the cage and new GpglTextureLayer for the slices.

The GpglTextureLayer is the implementation to display property values on triangles using a shader to define the colors.
The triangulated surface is expected to be "small", avoiding the need for a webworker.
It is not a private layer and can be uses as is.

Longer term design:
- move existing private layers, in particular PrivateTriangleLayer, PolylinesLayer, PrivatePointsLayer and PrivateMapLayer,
  to Gpgl versions. These standalone layers should be usable as plain layers, but are intended to be used as "implementation"
  layers for CompositeLayer.
  BTW, Gpgl stands for General Purpose Graphic Layers and are serving as base "WebGL/GpuGL" layers.

The layers come with associated stories.

Detail:
- add shared src/layers/utils/colormapTools.ts file, defining:
  - ColormapProps: specification of a colormap as a function, a typed array of colors or a colormapName associated to a colormap list.
  - ColormapSetup: settings of the colormap (value range, clamp range, clamp colors, undefined value and undefined color)
  and helper functions
  - to create a colormap texture
  - to normalize colors (transform from 0 - 255 range to 0 - 1 range)
- add src/layers/gpglLayers/typeDefs.ts file, defining:
  - Material
  - Texture2D to specify the value array

Move smooth form Texture2D to ColormapSetup