chore: "Typed Array Input" triangle story crash when editing layer settings (#2564)
This is a known issue of storybook. Story controls are handled as Json,
with stringified version used in the control panel. When edited, the
modified string version is transformed to a Json structure and send as
arguments to render the story.

This breaks the non Json compatible data, like arrays.

The triangle story was not using the workaround used to avoid crashes.