fix: draw unfolded well trajectories adjacent (#2623)
Wells are projected adjacent with a gap equal to the lateral distance
between the well trajectories' respective heads and ends.

<img width="937" height="332" alt="image"
src="https://github.com/user-attachments/assets/389a9ad6-d569-4eda-93bd-fc87f0057d5c"
/>


This is an improvement to #2019. Ideally projections should be more
configurable, but this is an improved default projection.

A follow-up improvement could be to expose the projection method as a
callback prop, using the current method as default. By using a custom
callback to define the projection, we could cover use cases such as
* partial trajectories (eg. you only want to use parts of respective
trajectories)
* arbitrary gaps (eg. through polyline digitization)
* reverse direction (eg. if we want the shortest path through the wells)